packages/ai/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ML and AI libraries
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4

# NLP libraries
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9

# Redis and caching
redis==5.0.1

# HTTP client
httpx==0.25.2

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Background processing
celery==5.3.4

# System monitoring
psutil==5.9.6

# Additional utilities
pillow==10.1.0
email-validator==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Database (optional)
sqlalchemy==2.0.23
asyncpg==0.29.0

# Logging and monitoring
structlog==23.2.0